Move download scheduler to services
